api/src/main/java/com/example/api/service/ApplyService.java
package com.example.api.service;

import com.example.api.producer.CouponCreateProducer;
import com.example.api.repository.AppliedUserRepository;
import com.example.api.repository.CouponCountRepository;
import com.example.api.repository.CouponRepository;
import org.springframework.stereotype.Service;

@Service
public class ApplyService {

    private final CouponRepository couponRepository;

    private final CouponCountRepository couponCountRepository;

    private final CouponCreateProducer couponCreateProducer;
    private final AppliedUserRepository appliedUserRepository;

    public ApplyService(CouponRepository couponRepository, CouponCountRepository couponCountRepository, CouponCreateProducer couponCreateProducer, AppliedUserRepository appliedUserRepository){
        this.couponRepository = couponRepository;
        this.couponCountRepository = couponCountRepository;
        this.couponCreateProducer = couponCreateProducer;
        this.appliedUserRepository = appliedUserRepository;
    }

    public void apply(Long userId){
        Long apply = appliedUserRepository.add(userId);

        if (apply != 1) {
            return;
        }

        // redis -> single thread -> race condition 해결 가능
        long count = couponCountRepository.increment(); // 쿠폰 개수 -> 핵심키 -> 정합성만 유지

        if (count > 100){
            return;
        }

        couponCreateProducer.create(userId);
    }
}
